requirements.txt
Flask
Flask-CORS
python-dotenv
gunicorn
psycopg2-binary
